+++ 5bit/index.js
@@ -1,0 +1,4 @@
+const bytesAsWords = require('./bytes_as_words');
+const wordsAsBytes = require('./words_as_bytes');
+
+module.exports = {bytesAsWords, wordsAsBytes};