+++ addresses/index.js
@@ -1,4 +1,9 @@
 const decodeBase58Address = require('./decode_base58_address');
 const decodeBech32Address = require('./decode_bech32_address');
+const encodeBech32Address = require('./encode_bech32_address');
 
-module.exports = {decodeBase58Address, decodeBech32Address};
+module.exports = {
+  decodeBase58Address,
+  decodeBech32Address,
+  encodeBech32Address,
+};