+++ hashes/index.js
@@ -1,8 +1,10 @@
+const hashForP2wpkh = require('./hash_for_p2wpkh');
 const idForBlock = require('./id_for_block');
 const idForTransactionComponents = require('./id_for_transaction_components');
 const previousBlockId = require('./previous_block_id');
 
 module.exports = {
+  hashForP2wpkh,
   idForBlock,
   idForTransactionComponents,
   previousBlockId,