+++ test/addresses/test_decode_bech32_address.js
@@ -90,7 +90,7 @@ const tests = [
       address: 'bc1p0xlxvlhemja6c4dqv22uapctqupfhlxm9h8z3k2e72q4k9hcz7v07qwwzcrf',
     },
     description: 'zero padding of more than 4 bits',
-    error: 'ExpectedNoPaddingOnBech32Address',
+    error: 'ExpectedNoPaddingWordsToConvertToDataBytes',
   },
   {
     args: {